+++ asgl/weights.py
@@ -50,7 +50,7 @@ class WEIGHTS:
     # PREPROCESSING ###################################################################################################
 
     def _preprocessing(self, power_weight):
-        if isinstance(power_weight, (np.int, np.float)):
+        if isinstance(power_weight, (int, float)):
             power_weight = [power_weight]
         return power_weight
